Refactor the plugin to be a datasource plugin
