+++ pkg/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"os"
 
-	"github.com/grafana/grafana-plugin-sdk-go/backend/app"
+	"github.com/grafana/grafana-plugin-sdk-go/backend/datasource"
 	"github.com/grafana/grafana-plugin-sdk-go/backend/log"
 
 	"github.com/grafana/synthetic-monitoring-app/pkg/plugin"
@@ -11,8 +11,6 @@ import (
 
 // version is the version of the plugin. It is set at build time by the Mage
 // target, which passes the version to the Go linker using `-X main.version=x.y.z`.
-const pluginId = "grafana-synthetic-monitoring-app"
-
 var version = "development"
 
 func main() {
@@ -20,9 +18,10 @@ func main() {
 
 	log.DefaultLogger.Info("Starting plugin process", "version", version)
 
-	// Manage handles the lifecycle of app instances, creating one per plugin
-	// context. This call blocks until Grafana shuts the process down.
-	if err := app.Manage(pluginId, plugin.NewApp, app.ManageOpts{}); err != nil {
+	// Manage handles the lifecycle of datasource instances, creating one per
+	// configured datasource. This call blocks until Grafana shuts the process
+	// down.
+	if err := datasource.Manage(plugin.ID, plugin.NewDatasource, datasource.ManageOpts{}); err != nil {
 		log.DefaultLogger.Error(err.Error())
 		os.Exit(1)
 	}
