+++ scripts/validate-backend
@@ -5,6 +5,9 @@
 # Grafana with dist/ mounted as the plugin directory, and asserts that Grafana
 # registered the plugin and that CheckHealth responds over gRPC.
 #
+# The backend is the *nested* datasource plugin (src/datasource/plugin.json), not
+# the app itself, so this provisions a datasource instance to check the health of.
+#
 # Runs in an isolated compose project with its own container name and port, so
 # it does not disturb a `yarn server` instance from this or any other checkout.
 
@@ -13,17 +16,34 @@ set -eu
 PORT=${PORT:-3199}
 PROJECT=sm-backend-validate
 NAME=grafana-sm-backend-validate
-PLUGIN_ID=grafana-synthetic-monitoring-app
+APP_ID=grafana-synthetic-monitoring-app
+PLUGIN_ID=synthetic-monitoring-datasource
+DS_UID=sm-backend-validate
 
 toplevel=$(git rev-parse --show-toplevel)
 cd "$toplevel"
 
 override=$(mktemp)
+provisioning=$(mktemp -d)
 cleanup() {
 	docker compose -f docker-compose.yaml -f "$override" -p "$PROJECT" down >/dev/null 2>&1 || true
 	rm -f "$override"
+	rm -rf "$provisioning"
 }
 trap cleanup EXIT INT TERM
+
+# Grafana only instantiates a datasource backend for a configured datasource, so
+# provision one. Mounted as a single file *inside* the dev/provisioning bind so
+# the checked-in provisioning is left untouched.
+cat > "$provisioning/sm-backend-validate.yaml" <<EOT
+apiVersion: 1
+datasources:
+  - name: SM Backend Validate
+    type: $PLUGIN_ID
+    uid: $DS_UID
+    access: proxy
+    isDefault: false
+EOT
 
 cat > "$override" <<EOT
 services:
@@ -31,6 +51,8 @@ services:
     container_name: $NAME
     ports: !override
       - $PORT:3000
+    volumes:
+      - $provisioning/sm-backend-validate.yaml:/etc/grafana/provisioning/datasources/sm-backend-validate.yaml
 EOT
 
 fail() {
@@ -50,8 +72,8 @@ fi
 echo "==> checking dist/ has a binary for linux (the container's OS)"
 arch=$(docker version --format '{{.Server.Arch}}')
 case "$arch" in
-	amd64) want=dist/gpx_sm_app_linux_amd64 ;;
-	arm64) want=dist/gpx_sm_app_linux_arm64 ;;
+	amd64) want=dist/datasource/gpx_sm_datasource_linux_amd64 ;;
+	arm64) want=dist/datasource/gpx_sm_datasource_linux_arm64 ;;
 	*) echo "FAIL: unsupported docker server arch: $arch" >&2; exit 1 ;;
 esac
 test -x "$want" || fail "missing executable $want"
@@ -60,11 +82,11 @@ echo "    found $want"
 echo "==> starting Grafana on :$PORT"
 docker compose -f docker-compose.yaml -f "$override" -p "$PROJECT" up --build -d >/dev/null
 
-echo "==> waiting for plugin health endpoint"
+echo "==> waiting for datasource health endpoint"
 health=""
 i=0
 while [ "$i" -lt 60 ]; do
-	health=$(curl -s -u admin:admin "http://localhost:$PORT/api/plugins/$PLUGIN_ID/health" 2>/dev/null || true)
+	health=$(curl -s -u admin:admin "http://localhost:$PORT/api/datasources/uid/$DS_UID/health" 2>/dev/null || true)
 	case "$health" in
 		*'"status":"OK"'*) break ;;
 	esac
@@ -83,7 +105,9 @@ case "$health" in
 	*) fail "health check did not return OK (got: ${health:-<no response>})" ;;
 esac
 
-echo "==> asserting Grafana registered the plugin"
+echo "==> asserting Grafana registered the app and its nested datasource"
+docker logs "$NAME" 2>&1 | grep -q "msg=\"Plugin registered\" pluginId=$APP_ID" \
+	|| fail "Grafana did not log 'Plugin registered' for $APP_ID"
 docker logs "$NAME" 2>&1 | grep -q "msg=\"Plugin registered\" pluginId=$PLUGIN_ID" \
 	|| fail "Grafana did not log 'Plugin registered' for $PLUGIN_ID"
 
@@ -93,8 +117,8 @@ if docker logs "$NAME" 2>&1 | grep -q "Could not start plugin backend"; then
 fi
 
 echo "==> asserting the plugin process is running in the container"
-docker exec "$NAME" ps ax 2>/dev/null | grep -q "gpx_sm_app_linux_" \
-	|| fail "no gpx_sm_app process found inside the container"
+docker exec "$NAME" ps ax 2>/dev/null | grep -q "gpx_sm_datasource_linux_" \
+	|| fail "no gpx_sm_datasource process found inside the container"
 
 echo "==> asserting the version ldflag was injected"
 # Read the working tree's package.json, not HEAD: the SDK's build target reads
